+++ lib/apidocs.js
@@ -7,6 +7,7 @@ module.exports = {
   friends: fs.readFileSync(path.join(__dirname, '../plugins/friends.md'), 'utf-8'),
   gossip: fs.readFileSync(path.join(__dirname, '../plugins/gossip.md'), 'utf-8'),
   invite: fs.readFileSync(path.join(__dirname, '../plugins/invite.md'), 'utf-8'),
+  plugins: fs.readFileSync(path.join(__dirname, '../plugins/plugins.md'), 'utf-8'),
   'private': fs.readFileSync(path.join(__dirname, '../plugins/private.md'), 'utf-8'),
   replicate: fs.readFileSync(path.join(__dirname, '../plugins/replicate.md'), 'utf-8')
 }
